(Refactor)[Hiero]: Extract hiero ui
